Segregate Blockhost handler
