+++ app/api/controllers.go
@@ -142,32 +142,6 @@ type blockHostRequest struct {
 	Host string `json:"host" binding:"required"` // Host can be IP or URL
 }
 
-func (api *Api) BlockHost(c *gin.Context) {
-	host := blockHostRequest{}
-	if err := c.BindJSON(&host); err != nil {
-		c.JSON(400, gin.H{"data": "error"})
-		c.AbortWithStatus(http.StatusBadRequest)
-		return
-	}
-
-	blockedHost, err := api.BlockHostUseCase.Block(host.Host)
-	if err != nil {
-		c.AbortWithStatus(http.StatusBadRequest)
-		c.JSON(http.StatusBadRequest, err)
-		return
-	}
-
-	if blockedHost == nil {
-		c.JSON(400, gin.H{"data": "Host not found"})
-		c.AbortWithStatus(http.StatusBadRequest)
-		return
-	}
-
-	c.Header("Access-Control-Allow-Origin", "*") //There is a vuln here, that's only for testing purpose.
-	c.Header("Access-Control-Allow-Methods", "POST")
-	c.JSON(http.StatusOK, gin.H{"message": "Host " + host.Host + "has been blocked"})
-}
-
 func (api *Api) SendAlertNotification(c *gin.Context) {
 	err := api.AlertsSender.SendLastAlertMessages()
 	if err != nil {
